requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.1
certifi==2016.9.26
cycler==0.10.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-registration==2.2
django-toolbelt==0.0.1
google-api-python-client==1.5.0
gunicorn==19.7.1
httplib2==0.9.2
lxml==3.6.4
matplotlib==1.5.3
numpy==1.11.2
oauth2client==2.0.2
packaging==16.8
panda==0.3.1
pandas==0.20.1
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.5.3
pytz==2017.2
requests==2.11.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.1.1
static3==0.7.0
stemming==1.0.1
uritemplate==0.6
urllib3==1.18
virtualenv==15.1.0
whitenoise==3.3.0
XlsxWriter==0.9.6
